#199 - Search media files
